add macOS power events
This brings in macOS power events.  Now one can listen for when the
system is going to sleep (waking up is a feature for another day).  This
introduces native bindings and a host of adjustments.  An example Sytter
has been added.

Earmarked with these changes is `rustfmt` applied to the code base.